backend/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
bcrypt==4.2.0
billiard==4.2.0
blinker==1.7.0
cachelib==0.9.0
celery==5.4.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.0
Flask-Caching==2.3.0
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.47.2
greenlet==3.0.3
gTTS==2.5.1
idna==3.6
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
kombu==5.4.0
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pandas==2.1.4
passlib==1.7.4
pillow==10.2.0
prompt_toolkit==3.0.47
pydub==0.25.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.8
reportlab==4.1.0
requests==2.31.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.1.2
